[vis] Support arbitrary line plots from Tensors
* Move tensor_as_field to vis_base
